Fix Windows format errors in dap_http_client.c
- DWORD format: use %lu with cast to unsigned long for buf_in_size

+++ module/net/server/http_server/http_client/dap_http_client.c
@@ -358,7 +358,7 @@ void dap_http_client_read( dap_events_socket_t *a_esocket, void *a_arg )
                         l_peol = NULL;
 
                 if ( !l_peol ) {
-                    log_it( L_ERROR, "Start-line with size %zu is not terminated by CRLF pair", a_esocket->buf_in_size);
+                    log_it( L_ERROR, "Start-line with size %lu is not terminated by CRLF pair", (unsigned long)a_esocket->buf_in_size);
                     s_report_error_and_restart( a_esocket, l_http_client, DAP_HTTP_STATUS_BAD_REQUEST );
                     break;
                 }
@@ -451,7 +451,7 @@ void dap_http_client_read( dap_events_socket_t *a_esocket, void *a_arg )
                         debug_if(s_debug_http, L_DEBUG, "May be incomplete request in buffer, wait another part");
                         return;
                     }
-                    log_it( L_ERROR, "Line with size %zu is not terminated by CRLF pair: %s", a_esocket->buf_in_size, a_esocket->buf_in);
+                    log_it( L_ERROR, "Line with size %lu is not terminated by CRLF pair: %s", (unsigned long)a_esocket->buf_in_size, a_esocket->buf_in);
                     s_report_error_and_restart( a_esocket, l_http_client, DAP_HTTP_STATUS_BAD_REQUEST );
                     break;
                 }
